# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "XXC-APTD management",
    "version": "0.1.0",
    "description": "Unix socket: filesystem administrator authority. Administrative HTTP: local session, RBAC, exact Origin and CSRF for mutations. Public HTTP never serves these endpoints. Scoped project bearer tokens authorize only explicit publishing routes and suites. /admin is an equivalent private TCP API prefix."
  },
  "servers": [
    {
      "url": "https://apt.thugs.red/admin",
      "description": "Example path proxy. Confirm the administrative API origin with the operator; public /api is documentation only."
    },
    {
      "url": "https://admin.example.invalid",
      "description": "Example dedicated admin origin; replace with the operator-provided value. Paths below already include /api/v1."
    }
  ],
  "paths": {
    "/api/v1/status": {
      "get": {
        "summary": "status",
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "x-minimum-role": "viewer",
        "security": [
          {
            "projectToken": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-project-scope": "read"
      },
      "parameters": [
        {
          "name": "suite",
          "in": "query",
          "schema": {
            "type": "string"
          },
          "description": "Configured suite; default repository.suite. Required when token lacks access to default."
        }
      ]
    },
...
    "/api/v1/packages": {
      "get": {
        "summary": "packages",
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "x-minimum-role": "viewer",
        "security": [
          {
            "projectToken": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-project-scope": "read"
      },
      "parameters": [
        {
          "name": "suite",
          "in": "query",
          "schema": {
            "type": "string"
          },
          "description": "Configured suite; default repository.suite. Required when token lacks access to default."
        }
      ]
    },
    "/api/v1/packages/{id}": {
      "get": {
        "summary": "packages by ID",
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Package"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-minimum-role": "viewer",
        "security": [
          {
            "projectToken": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-project-scope": "read"
      },
      "parameters": [
        {
          "name": "suite",
          "in": "query",
          "schema": {
            "type": "string"
          },
          "description": "Configured suite; default repository.suite. Required when token lacks access to default."
        }
      ]
    },
    "/api/v1/uploads": {
      "get": {
        "summary": "uploads",
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "x-minimum-role": "viewer",
        "security": [
          {
            "projectToken": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-project-scope": "read"
      },
      "post": {
        "summary": "uploads",
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Package"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "408": {
            "$ref": "#/components/responses/Error408"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "application/vnd.debian.binary-package": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/Origin"
          },
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ],
        "x-minimum-role": "operator",
        "security": [
          {
            "projectToken": []
          },
          {
            "SessionCookie": []
          }
        ],
        "x-project-scope": "upload"
      },
      "parameters": [
        {
          "name": "suite",
          "in": "query",
          "schema": {
            "type": "string"
          },
          "description": "Configured suite; default repository.suite. Required when token lacks access to default."
        }
      ]
    },
    "/api/v1/repository/generations": {
      "get": {
        "summary": "repository generations",
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "x-minimum-role": "viewer"
      }
    },
...
    "/api/v1/audit": {
      "get": {
        "summary": "audit",
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          }
        },
        "x-minimum-role": "viewer"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-xxc-session",
        "description": "Admin HTTP interface only. When admin.external_url uses HTTPS, the cookie is __Host-xxc-session with Secure. Explicit HTTP administration uses xxc-session, independently of the public origin or proxy headers. Socket requests authenticate through filesystem permissions."
      },
      "projectToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "One-time project token with explicit scope and suite allowlists"
      }
    },
...
    "headers": {
      "RequestId": {
        "schema": {
          "type": "string"
        },
        "description": "Correlation identifier"
      },
      "NoStore": {
        "schema": {
          "type": "string",
          "enum": [
            "no-store"
          ]
        }
      }
    },
    "responses": {
...
      "Error404": {
        "description": "Unknown object",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          },
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "SessionCookie": []
    }
  ],
...
}